ft. added a CPP Tree solution
Q: Lowest Common Ancestor in a Binary Tree
Given a Binary Tree with all unique values and two nodes value, n1 and n2. The task is to find the lowest common ancestor of the given two nodes. We may assume that either both n1 and n2 are present in the tree or none of them are present.

LCA: It is the first common ancestor of both the nodes n1 and n2 from bottom of tree.

Solution of the above is done in CPP in the CPP folder, in folder titled: "Lowest Common Ancestor of Binary Tree"