added a famous Medium level Tree qna
Q:Given a binary tree of size N. Your task is to complete the function sumOfLongRootToLeafPath(), that find the sum of all nodes on the longest path from root to leaf node.
If two or more paths compete for the longest path, then the path having maximum sum of nodes is being considered.

I've given the solution to the following problem in the CPP folder in a folder titled: Longest Max Sum Bloodline of Binary Tree